As suggested in the code review, refactor StartTransaction and create a new file for the APM specific
