+++ e2e/templates/suite_test.go.tmpl
@@ -60,7 +60,7 @@ func Initialize${CAPITAL_SUITE}Scenarios(ctx *godog.ScenarioContext) {
 	ctx.BeforeScenario(func(p *messages.Pickle) {
 		log.Trace("Before ${CAPITAL_SUITE} scenario...")
 
-		tx = utils.StartTransaction(p.GetName(), "test.scenario")
+		tx = apme2e.StartTransaction(p.GetName(), "test.scenario")
 		tx.Context.SetLabel("suite", "${CAPITAL_SUITE}")
 	})
 
@@ -113,7 +113,7 @@ func Initialize${CAPITAL_SUITE}TestSuite(ctx *godog.TestSuiteContext) {
 
 		// instrumentation
 		defer apm.DefaultTracer.Flush(nil)
-		suiteTx = utils.StartTransaction("Initialise ${CAPITAL_SUITE}", "test.suite")
+		suiteTx = apme2e.StartTransaction("Initialise ${CAPITAL_SUITE}", "test.suite")
 		defer suiteTx.End()
 		suiteParentSpan = suiteTx.StartSpan("Before ${CAPITAL_SUITE} test suite", "test.suite.before", nil)
 		suiteContext = apm.ContextWithSpan(suiteContext, suiteParentSpan)
@@ -134,7 +134,7 @@ func Initialize${CAPITAL_SUITE}TestSuite(ctx *godog.TestSuiteContext) {
 		var suiteParentSpan *apm.Span
 		var suiteContext = context.Background()
 		defer apm.DefaultTracer.Flush(nil)
-		suiteTx = utils.StartTransaction("Tear Down ${CAPITAL_SUITE}", "test.suite")
+		suiteTx = apme2e.StartTransaction("Tear Down ${CAPITAL_SUITE}", "test.suite")
 		defer suiteTx.End()
 		suiteParentSpan = suiteTx.StartSpan("After ${CAPITAL_SUITE} test suite", "test.suite.after", nil)
 		suiteContext = apm.ContextWithSpan(suiteContext, suiteParentSpan)
