+++ internal/utils/utils.go
@@ -684,21 +684,3 @@ func GetDockerNamespaceEnvVar(fallback string) string {
 	}
 	return fallback
 }
-
-// StartTransaction returns a new Transaction with the specified
-// name and type, with the start time set to the current time and
-// with the context if TRACEPARENT environment variable is set.
-// This is equivalent to calling apm.DefaultTracer.StartTransaction
-// if no TRACEPARENT environment variable otherwise
-// apm.DefaultTracer.StartTransactionOptions
-func StartTransaction(name, transactionType string) *apm.Transaction {
-	traceparent := shell.GetEnv("TRACEPARENT", "")
-	if traceparent != "" {
-		log.WithFields(log.Fields{
-			"traceparent": traceparent,
-		}).Debug("Using the given traceparent")
-		return apm.DefaultTracer.StartTransaction(name, transactionType)
-	}
-
-	return apm.DefaultTracer.StartTransaction(name, transactionType)
-}
